samples/python/agentic_ai/research_organizer/requirements.txt
rich>=13.0.0
requests>=2.28.0
fastmcp>=0.1.0
